Set tactic completions when initializing editor. (#195)
* Set tactic completions when initializing editor.

* Clean-up as per review

* Remove syntax message
