+++ editor/src/index.ts
@@ -3,6 +3,8 @@ import { WaterproofEditor, WaterproofEditorConfig } from "./waterproof-editor";
 // TODO: Move this to a types location.
 import { TextDocMappingMV, TextDocMappingV } from "./mapping";
 import { blocksFromMV, blocksFromV } from "./document-construction/construct-document";
+// TODO: The tactics completions are static, we want them to be dynamic (LSP supplied and/or configurable when the editor is running)
+import tactics from "../../shared/completions/tactics.json";
 
 /**
  * Very basic representation of the acquirable VSCodeApi.
@@ -15,6 +17,7 @@ interface VSCodeAPI {
 function createConfiguration(format: FileFormat, codeAPI: VSCodeAPI) {
 	// Create the WaterproofEditorConfig object
 	const cfg: WaterproofEditorConfig = {
+		completions: tactics,
 		api: {
 			executeHelp() {
 				codeAPI.postMessage({ type: MessageType.command, body: { command: "Help.", time: (new Date()).getTime()} });
@@ -117,9 +120,6 @@ window.onload = () => {
 				{ const diagnostics = msg.body;
 				editor.parseCoqDiagnostics(diagnostics);
 				break; }
-			case MessageType.syntax:
-				editor.initTacticCompletion(msg.body);
-				break;
 			default:
 				// If we reach this 'default' case, then we have encountered an unknown message type.
 				console.log(`[WEBVIEW] Unrecognized message type '${msg.type}'`);
