+++ editor/src/waterproof-editor/autocomplete/index.ts
@@ -2,5 +2,4 @@
 export { emojiCompletionSource } from "./emojis";
 export { coqCompletionSource } from "./coqTerms";
 export { symbolCompletionSource } from "./symbols";
-export { tacticCompletionSource } from "./tactics";
 export { renderIcon } from "./renderSymbol";
