+++ shared/Messages.ts
@@ -46,7 +46,6 @@ export type Message =
     | MessageBase<MessageType.setData, string[] | GoalAnswer<PpString> >
     | MessageBase<MessageType.setShowLineNumbers, boolean>
     | MessageBase<MessageType.setShowMenuItems, boolean>
-    | MessageBase<MessageType.syntax, boolean>
     | MessageBase<MessageType.teacher, boolean>;
 
 /**
@@ -75,7 +74,6 @@ export const enum MessageType {
     setData,
     setShowLineNumbers,
     setShowMenuItems,
-    syntax,
     teacher,
 }
